+++ PlayTogether/PlayTogether/Sources/Chatting/Cell/ChattingListTableViewCell.swift
@@ -106,13 +106,4 @@ private extension ChattingListTableViewCell {
             $0.trailing.equalToSuperview().inset(20)
         }
     }
-    
-    func dateParser(_ dateString: String) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
-        guard let convertDate = dateFormatter.date(from: dateString) else { return String.init() }
-        
-        dateFormatter.dateFormat = "yyyy.MM.dd.  HH:mm"
-        return dateFormatter.string(from: convertDate)
-    }
 }
